fix c compiler float printing

+++ tests/machine/x/c/python_math.c
@@ -11,14 +11,14 @@ extern double math_log(double x);
 static double r = 3;
 
 int main() {
-  int area = 3.141592653589793 * (r * r);
+  double area = 3.141592653589793 * (r * r);
   double root = 7;
   double sin45 = __builtin_sin(3.141592653589793 / 4.0);
   double log_e = __builtin_log(2.718281828459045);
   printf("%s ", "Circle area with r =");
-  printf("%d ", r);
+  printf("%.16g ", r);
   printf("%s ", "=>");
-  printf("%d\n", area);
+  printf("%.16g\n", area);
   printf("%s ", "Square root of 49:");
   printf("%.16g\n", root);
   printf("%s ", "sin(π/4):");
